+++ moto/config/models.py
@@ -45,7 +45,7 @@ from moto.config.exceptions import (
 from moto.core import BaseBackend, BaseModel
 from moto.s3.config import s3_config_query
 
-from moto.iam.models import ACCOUNT_ID as DEFAULT_ACCOUNT_ID
+from moto.core import ACCOUNT_ID as DEFAULT_ACCOUNT_ID
 POP_STRINGS = [
     "capitalizeStart",
     "CapitalizeStart",
