+++ moto/core/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import unicode_literals
 
-from .models import BaseModel, BaseBackend, moto_api_backend  # noqa
+from .models import BaseModel, BaseBackend, moto_api_backend, ACCOUNT_ID  # noqa
 from .responses import ActionAuthenticatorMixin
 
 moto_api_backends = {"global": moto_api_backend}
