+++ moto/core/access_control.py
@@ -24,7 +24,8 @@ from botocore.awsrequest import AWSRequest
 from botocore.credentials import Credentials
 from six import string_types
 
-from moto.iam.models import ACCOUNT_ID, Policy
+from moto.core import ACCOUNT_ID 
+from moto.iam.models import Policy
 from moto.iam import iam_backend
 from moto.core.exceptions import (
     SignatureDoesNotMatchError,
