+++ moto/ec2/responses/security_groups.py
@@ -2,7 +2,7 @@ from __future__ import unicode_literals
 
 from moto.core.responses import BaseResponse
 from moto.ec2.utils import filters_from_querystring
-from moto.iam.models import ACCOUNT_ID
+from moto.core import ACCOUNT_ID
 
 
 def try_parse_int(value, default=None):
