+++ moto/ec2/responses/vpc_peering_connections.py
@@ -1,6 +1,6 @@
 from __future__ import unicode_literals
 from moto.core.responses import BaseResponse
-from moto.iam.models import ACCOUNT_ID
+from moto.core import ACCOUNT_ID
 
 
 class VPCPeeringConnections(BaseResponse):
