+++ moto/redshift/models.py
@@ -27,7 +27,7 @@ from .exceptions import (
 )
 
 
-from moto.iam.models import ACCOUNT_ID
+from moto.core import ACCOUNT_ID
 
 
 class TaggableResourceMixin(object):
