+++ tests/test_cloudformation/test_cloudformation_stack_crud.py
@@ -14,7 +14,7 @@ import sure  # noqa
 # Ensure 'assert_raises' context manager support for Python 2.6
 import tests.backport_assert_raises  # noqa
 from nose.tools import assert_raises
-from moto.iam.models import ACCOUNT_ID
+from moto.core import ACCOUNT_ID
 
 from moto import (
     mock_cloudformation_deprecated,
