+++ tests/test_config/test_config.py
@@ -7,7 +7,7 @@ from nose.tools import assert_raises
 
 from moto import mock_s3
 from moto.config import mock_config
-from moto.iam.models import ACCOUNT_ID
+from moto.core import ACCOUNT_ID
 
 @mock_config
 def test_put_configuration_recorder():
