+++ tests/test_kinesis/test_firehose.py
@@ -7,7 +7,7 @@ import boto3
 import sure  # noqa
 
 from moto import mock_kinesis
-from moto.iam.models import ACCOUNT_ID
+from moto.core import ACCOUNT_ID
 
 
 def create_s3_delivery_stream(client, stream_name):
