+++ tests/test_ses/test_ses_sns_boto3.py
@@ -10,7 +10,7 @@ import sure  # noqa
 from nose import tools
 from moto import mock_ses, mock_sns, mock_sqs
 from moto.ses.models import SESFeedback
-from moto.iam.models import ACCOUNT_ID
+from moto.core import ACCOUNT_ID
 
 
 @mock_ses
