+++ tests/test_sns/test_application_boto3.py
@@ -4,7 +4,7 @@ import boto3
 from botocore.exceptions import ClientError
 from moto import mock_sns
 import sure  # noqa
-from moto.iam.models import ACCOUNT_ID
+from moto.core import ACCOUNT_ID
 
 
 @mock_sns
